+++ packages/network-enablement-controller/src/NetworkEnablementController.ts
@@ -117,9 +117,13 @@ const getDefaultNetworkEnablementControllerState =
       },
       [KnownCaipNamespace.Solana]: {
         [SolScope.Mainnet]: true,
+        [SolScope.Testnet]: false,
+        [SolScope.Devnet]: false,
       },
       [KnownCaipNamespace.Bip122]: {
         [BtcScope.Mainnet]: true,
+        [BtcScope.Testnet]: false,
+        [BtcScope.Signet]: false,
       },
     },
   });
@@ -177,18 +181,6 @@ export class NetworkEnablementController extends BaseController<
     messenger.subscribe('NetworkController:networkRemoved', ({ chainId }) => {
       this.#removeNetworkEntry(chainId);
     });
-
-    // Listen for confirmed staking transactions
-    messenger.subscribe(
-      'TransactionController:transactionSubmitted',
-      (transactionMeta) => {
-        if (transactionMeta?.transactionMeta?.chainId) {
-          this.enableNetwork(
-            transactionMeta.transactionMeta.chainId as Hex | CaipChainId,
-          );
-        }
-      },
-    );
   }
 
   /**
@@ -384,10 +376,6 @@ export class NetworkEnablementController extends BaseController<
   disableNetwork(chainId: Hex | CaipChainId): void {
     const derivedKeys = deriveKeys(chainId);
     const { namespace, storageKey } = derivedKeys;
-
-    if (isOnlyNetworkEnabledInNamespace(this.state, derivedKeys)) {
-      throw new Error('Cannot disable the last remaining enabled network');
-    }
 
     this.update((s) => {
       s.enabledNetworkMap[namespace][storageKey] = false;