+++ packages/subscription-controller/src/SubscriptionController.ts
@@ -11,7 +11,15 @@ import {
   controllerName,
   SubscriptionControllerErrorMessage,
 } from './constants';
+import type {
+  GetCryptoApproveTransactionRequest,
+  GetCryptoApproveTransactionResponse,
+  ProductPrice,
+  StartCryptoSubscriptionRequest,
+  TokenPaymentInfo,
+} from './types';
 import {
+  PaymentType,
   SubscriptionStatus,
   type ISubscriptionService,
   type PricingResponse,
@@ -41,6 +49,14 @@ export type SubscriptionControllerGetPricingAction = {
   type: `${typeof controllerName}:getPricing`;
   handler: SubscriptionController['getPricing'];
 };
+export type SubscriptionControllerGetCryptoApproveTransactionParamsAction = {
+  type: `${typeof controllerName}:getCryptoApproveTransactionParams`;
+  handler: SubscriptionController['getCryptoApproveTransactionParams'];
+};
+export type SubscriptionControllerStartSubscriptionWithCryptoAction = {
+  type: `${typeof controllerName}:startSubscriptionWithCrypto`;
+  handler: SubscriptionController['startSubscriptionWithCrypto'];
+};
 
 export type SubscriptionControllerGetStateAction = ControllerGetStateAction<
   typeof controllerName,
@@ -51,10 +67,13 @@ export type SubscriptionControllerActions =
   | SubscriptionControllerCancelSubscriptionAction
   | SubscriptionControllerStartShieldSubscriptionWithCardAction
   | SubscriptionControllerGetPricingAction
-  | SubscriptionControllerGetStateAction;
+  | SubscriptionControllerGetStateAction
+  | SubscriptionControllerGetCryptoApproveTransactionParamsAction
+  | SubscriptionControllerStartSubscriptionWithCryptoAction;
 
 export type AllowedActions =
-  AuthenticationController.AuthenticationControllerGetBearerToken;
+  | AuthenticationController.AuthenticationControllerGetBearerToken
+  | AuthenticationController.AuthenticationControllerPerformSignOut;
 
 // Events
 export type SubscriptionControllerStateChangeEvent = ControllerStateChangeEvent<
@@ -150,7 +169,6 @@ export class SubscriptionController extends BaseController<
     });
 
     this.#subscriptionService = subscriptionService;
-
     this.#registerMessageHandlers();
   }
 
@@ -177,6 +195,16 @@ export class SubscriptionController extends BaseController<
     this.messagingSystem.registerActionHandler(
       'SubscriptionController:getPricing',
       this.getPricing.bind(this),
+    );
+
+    this.messagingSystem.registerActionHandler(
+      'SubscriptionController:getCryptoApproveTransactionParams',
+      this.getCryptoApproveTransactionParams.bind(this),
+    );
+
+    this.messagingSystem.registerActionHandler(
+      'SubscriptionController:startSubscriptionWithCrypto',
+      this.startSubscriptionWithCrypto.bind(this),
     );
   }
 
@@ -214,12 +242,135 @@ export class SubscriptionController extends BaseController<
           : subscription,
       );
     });
+
+    this.triggerAccessTokenRefresh();
   }
 
   async startShieldSubscriptionWithCard(request: StartSubscriptionRequest) {
     this.#assertIsUserNotSubscribed({ products: request.products });
 
-    return await this.#subscriptionService.startSubscriptionWithCard(request);
+    const response =
+      await this.#subscriptionService.startSubscriptionWithCard(request);
+
+    this.triggerAccessTokenRefresh();
+
+    return response;
+  }
+
+  async startSubscriptionWithCrypto(request: StartCryptoSubscriptionRequest) {
+    this.#assertIsUserNotSubscribed({ products: request.products });
+    return await this.#subscriptionService.startSubscriptionWithCrypto(request);
+  }
+
+  /**
+   * Get transaction params to create crypto approve transaction for subscription payment
+   *
+   * @param request - The request object
+   * @param request.chainId - The chain ID
+   * @param request.tokenAddress - The address of the token
+   * @param request.productType - The product type
+   * @param request.interval - The interval
+   * @returns The crypto approve transaction params
+   */
+  async getCryptoApproveTransactionParams(
+    request: GetCryptoApproveTransactionRequest,
+  ): Promise<GetCryptoApproveTransactionResponse> {
+    const pricing = await this.getPricing();
+    const product = pricing.products.find(
+      (p) => p.name === request.productType,
+    );
+    if (!product) {
+      throw new Error('Product price not found');
+    }
+
+    const price = product.prices.find((p) => p.interval === request.interval);
+    if (!price) {
+      throw new Error('Price not found');
+    }
+
+    const chainsPaymentInfo = pricing.paymentMethods.find(
+      (t) => t.type === PaymentType.byCrypto,
+    );
+    if (!chainsPaymentInfo) {
+      throw new Error('Chains payment info not found');
+    }
+    const chainPaymentInfo = chainsPaymentInfo.chains?.find(
+      (t) => t.chainId === request.chainId,
+    );
+    if (!chainPaymentInfo) {
+      throw new Error('Invalid chain id');
+    }
+    const tokenPaymentInfo = chainPaymentInfo.tokens.find(
+      (t) => t.address === request.paymentTokenAddress,
+    );
+    if (!tokenPaymentInfo) {
+      throw new Error('Invalid token address');
+    }
+
+    const tokenApproveAmount = this.#getTokenApproveAmount(
+      price,
+      tokenPaymentInfo,
+    );
+
+    return {
+      approveAmount: tokenApproveAmount.toString(),
+      paymentAddress: chainPaymentInfo.paymentAddress,
+      paymentTokenAddress: request.paymentTokenAddress,
+      chainId: request.chainId,
+    };
+  }
+
+  /**
+   * Calculate total subscription price amount from price info
+   * e.g: $8 per month * 12 months min billing cycles = $96
+   *
+   * @param price - The price info
+   * @returns The price amount
+   */
+  #getSubscriptionPriceAmount(price: ProductPrice) {
+    // no need to use BigInt since max unitDecimals are always 2 for price
+    const amount =
+      (price.unitAmount / 10 ** price.unitDecimals) * price.minBillingCycles;
+    return amount;
+  }
+
+  /**
+   * Calculate token approve amount from price info
+   *
+   * @param price - The price info
+   * @param tokenPaymentInfo - The token price info
+   * @returns The token approve amount
+   */
+  #getTokenApproveAmount(
+    price: ProductPrice,
+    tokenPaymentInfo: TokenPaymentInfo,
+  ) {
+    const conversionRate =
+      tokenPaymentInfo.conversionRate[
+        price.currency as keyof typeof tokenPaymentInfo.conversionRate
+      ];
+    if (!conversionRate) {
+      throw new Error('Conversion rate not found');
+    }
+    // conversion rate is a float string e.g: "1.0"
+    // We need to handle float conversion rates with integer math for BigInt.
+    // We'll scale the conversion rate to an integer by multiplying by 10^4.
+    // conversionRate is in usd decimal. In most currencies, we only care about 2 decimals (cents)
+    // So, scale must be max of 10 ** 4 (most exchanges trade with max 4 decimals of usd)
+    // This allows us to avoid floating point math and keep precision.
+    const SCALE = 10n ** 4n;
+    const conversionRateScaled =
+      BigInt(Math.round(Number(conversionRate) * Number(SCALE))) / SCALE;
+    // price of the product
+    const priceAmount = this.#getSubscriptionPriceAmount(price);
+    const priceAmountScaled =
+      BigInt(Math.round(priceAmount * Number(SCALE))) / SCALE;
+
+    const tokenDecimal = BigInt(10) ** BigInt(tokenPaymentInfo.decimals);
+
+    const tokenAmount =
+      (priceAmountScaled * tokenDecimal) / conversionRateScaled;
+    return tokenAmount;
   }
 
   #assertIsUserNotSubscribed({ products }: { products: ProductType[] }) {
@@ -230,6 +381,16 @@ export class SubscriptionController extends BaseController<
     ) {
       throw new Error(SubscriptionControllerErrorMessage.UserAlreadySubscribed);
     }
+  }
+
+  /**
+   * Triggers an access token refresh.
+   */
+  triggerAccessTokenRefresh() {
+    // We perform a sign out to clear the access token from the authentication
+    // controller. Next time the access token is requested, a new access token
+    // will be fetched.
+    this.messagingSystem.call('AuthenticationController:performSignOut');
   }
 
   #assertIsUserSubscribed(request: { subscriptionId: string }) {