+++ packages/subscription-controller/src/SubscriptionService.ts
@@ -1,3 +1,5 @@
+import { handleFetch } from '@metamask/controller-utils';
+
 import {
   getEnvUrls,
   SubscriptionControllerErrorMessage,
@@ -9,6 +11,8 @@ import type {
   GetSubscriptionsResponse,
   ISubscriptionService,
   PricingResponse,
+  StartCryptoSubscriptionRequest,
+  StartCryptoSubscriptionResponse,
   StartSubscriptionRequest,
   StartSubscriptionResponse,
 } from './types';
@@ -16,28 +20,19 @@ import type {
 export type SubscriptionServiceConfig = {
   env: Env;
   auth: AuthUtils;
-  fetchFn: typeof globalThis.fetch;
-};
-
-type ErrorMessage = {
-  message: string;
-  error: string;
 };
 
 export const SUBSCRIPTION_URL = (env: Env, path: string) =>
-  `${getEnvUrls(env).subscriptionApiUrl}/api/v1/${path}`;
+  `${getEnvUrls(env).subscriptionApiUrl}/v1/${path}`;
 
 export class SubscriptionService implements ISubscriptionService {
   readonly #env: Env;
-
-  readonly #fetch: typeof globalThis.fetch;
 
   public authUtils: AuthUtils;
 
   constructor(config: SubscriptionServiceConfig) {
     this.#env = config.env;
     this.authUtils = config.auth;
-    this.#fetch = config.fetchFn;
   }
 
   async getSubscriptions(): Promise<GetSubscriptionsResponse> {
@@ -63,6 +58,13 @@ export class SubscriptionService implements ISubscriptionService {
     return await this.#makeRequest(path, 'POST', request);
   }
 
+  async startSubscriptionWithCrypto(
+    request: StartCryptoSubscriptionRequest,
+  ): Promise<StartCryptoSubscriptionResponse> {
+    const path = 'subscriptions/crypto';
+    return await this.#makeRequest(path, 'POST', request);
+  }
+
   async #makeRequest<Result>(
     path: string,
     method: 'GET' | 'POST' | 'DELETE' | 'PUT' | 'PATCH' = 'GET',
@@ -72,7 +74,7 @@ export class SubscriptionService implements ISubscriptionService {
       const headers = await this.#getAuthorizationHeader();
       const url = new URL(SUBSCRIPTION_URL(this.#env, path));
 
-      const response = await this.#fetch(url.toString(), {
+      const response = await handleFetch(url.toString(), {
         method,
         headers: {
           'Content-Type': 'application/json',
@@ -81,16 +83,9 @@ export class SubscriptionService implements ISubscriptionService {
         body: body ? JSON.stringify(body) : undefined,
       });
 
-      const responseBody = await response.json();
-      if (!response.ok) {
-        const { message, error } = responseBody as ErrorMessage;
-        throw new Error(`HTTP error message: ${message}, error: ${error}`);
-      }
-
-      return responseBody as Result;
+      return response;
     } catch (e) {
-      const errorMessage =
-        e instanceof Error ? e.message : JSON.stringify(e ?? 'unknown error');
+      const errorMessage = e instanceof Error ? e.message : JSON.stringify(e);
 
       throw new SubscriptionServiceError(
         `failed to make request. ${errorMessage}`,